% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{a}
\alias{a}
\alias{na}
\title{Shortcuts for \code{\link[=is.na]{is.na()}}}
\usage{
a(x)

na(x)
}
\arguments{
\item{x}{A vector to be tested.}
}
\value{
A logical vector.
}
\description{
\code{a(x)} is the same as \code{!is.na(x)} and \code{na()} is an alias for \code{is.na()}.
}
\examples{
a(c(1, NA, 3))
na(c(1, NA, 3))

}
\seealso{
\code{\link[=is.na]{is.na()}}
}
