% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmf-package.R
\docType{package}
\name{jmf-package}
\alias{jmf}
\alias{jmf-package}
\title{jmf: JMF R Utility Functions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

JMF R helpers and common functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jmf-vienna.github.io/jmf-r/}
  \item Report bugs at \url{https://github.com/jmf-vienna/jmf-r/issues}
}

}
\author{
\strong{Maintainer}: Joint Microbiome Facility (JMF) of the Medical University of Vienna and the University of Vienna \email{info.jmf@univie.ac.at} [copyright holder, funder]

Authors:
\itemize{
  \item Bela Hausmann (\href{https://orcid.org/0000-0002-0846-1202}{ORCID})
}

}
\keyword{internal}
