% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniques.R
\name{uniques}
\alias{uniques}
\title{Unique and sorted vectors}
\usage{
uniques(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A sorted vector without duplicates.
}
\description{
Like \code{\link[base:unique]{unique()}}, but also sorted.
}
\examples{
base::unique(c(3, 1, 2, 3))
uniques(c(3, 1, 2, 3))

}
\seealso{
\code{\link[base:unique]{unique()}}
}
