% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-jmf-project-id.R
\name{is_jmf_project_id}
\alias{is_jmf_project_id}
\alias{jmf_project_id_regex}
\title{Check for valid JMF IDs}
\usage{
is_jmf_project_id(string, internal = "deny")

jmf_project_id_regex(internal = "deny")
}
\arguments{
\item{string}{Input vector passed to \link[stringr:str_detect]{str_detect()}.}

\item{internal}{\code{"allow"} or \code{"deny"} internal IDs.}
}
\value{
\itemize{
\item \code{is_jmf_project_id()}: a logical vector the same length as \code{string}.
\item \code{jmf_project_id_regex()}: a \link[rex:rex]{rex()} regular expression.
}
}
\description{
\itemize{
\item \code{is_jmf_project_id()}: checks for valid JMF project IDs.
\item \code{jmf_project_id_regex()}: returns the underlying regular expression.
}
}
\examples{
# valid JMF project ID:
is_jmf_project_id("JMF-2401-42")

# this is a JMF sample ID:
is_jmf_project_id("JMF-2401-42-0001")

# extract JMF project ID from text:
stringr::str_extract(
  "my sample id is JMF-2401-42-0001",
  jmf_project_id_regex()
)

}
